+++ packages/video/ios/GraniteVideoView.mm
@@ -244,8 +244,8 @@ using namespace facebook::react;
     }
 
     videoSource.startPosition = source.startPosition;
-    videoSource.cropStart = source.cropStart;
-    videoSource.cropEnd = source.cropEnd;
+    videoSource.startTime = source.startTime;
+    videoSource.endTime = source.endTime;
 
     [_provider loadSource:videoSource];
 
@@ -534,6 +534,17 @@ using namespace facebook::react;
     }
 }
 
+- (void)videoAudioFocusChangedWithHasAudioFocus:(BOOL)hasAudioFocus
+{
+    if (_eventEmitter) {
+        auto emitter = std::static_pointer_cast<GraniteVideoViewEventEmitter const>(_eventEmitter);
+        facebook::react::GraniteVideoViewEventEmitter::OnVideoAudioFocusChanged event = {
+            .hasAudioFocus = hasAudioFocus
+        };
+        emitter->onVideoAudioFocusChanged(event);
+    }
+}
+
 - (void)videoFullscreenPlayerWillPresent
 {
     if (_eventEmitter) {
@@ -644,6 +655,7 @@ using namespace facebook::react;
 {
     [super prepareForRecycle];
     [_provider unload];
+    _props = std::make_shared<const GraniteVideoViewProps>();
 }
 
 - (void)dealloc