+++ packages/jest/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @granite-js/jest
 
+## 1.0.12
+
+### Patch Changes
+
+- @granite-js/native@1.0.12
+
+## 1.0.11
+
+### Patch Changes
+
+- @granite-js/native@1.0.11
+
 ## 1.0.10
 
 ### Patch Changes