+++ packages/plugin-env/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @granite-js/plugin-env
 
+## 1.0.12
+
+### Patch Changes
+
+- @granite-js/plugin-core@1.0.12
+- @granite-js/utils@1.0.12
+
+## 1.0.11
+
+### Patch Changes
+
+- @granite-js/plugin-core@1.0.11
+- @granite-js/utils@1.0.11
+
 ## 1.0.10
 
 ### Patch Changes