+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @granite-js/react-native
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [61b8058]
+- Updated dependencies [61b8058]
+  - @granite-js/video@1.0.12
+  - @granite-js/native@1.0.12
+  - @granite-js/blur-view@1.0.12
+  - @granite-js/jest@1.0.12
+  - @granite-js/style-utils@1.0.12
+  - @granite-js/cli@1.0.12
+  - @granite-js/mpack@1.0.12
+  - @granite-js/plugin-core@1.0.12
+
+## 1.0.11
+
+### Patch Changes
+
+- @granite-js/native@1.0.11
+- @granite-js/blur-view@1.0.11
+- @granite-js/jest@1.0.11
+- @granite-js/style-utils@1.0.11
+- @granite-js/cli@1.0.11
+- @granite-js/mpack@1.0.11
+- @granite-js/plugin-core@1.0.11
+- @granite-js/video@1.0.11
+
 ## 1.0.10
 
 ### Patch Changes